Added files via upload
Now when a user logs back in it makes a player object and adds their data in
I also made a string to int function so that the pet number could always be represented as an int. I imagine this will be useful for the bank account as well.
My next goal will be to prompt the user to make a pet if they have zero pets and then load some info on game play.
I am thinking adding some instructions to the menu might be nice.