Added files via upload
lots of errors in this code right now... but basically a user can create an account, the program can load the user and the user can create one pet that the program can recognize the name of. I will eventually go through and make it so multiple pets can be accepted and more information can be loaded besides pet name. This is most likely tomorrow goal.


also some notes on the vector storage
0: username
1: petnumber
2: petname
3: pettype

pet number is the number of pets the user has and pet type is the numerical type the user selects.